refactor: Animal 패키지의 wrapper 패키지 이름을 vo로 변경한다.
