+++ src/main/java/com/clova/anifriends/domain/animal/dto/request/FindAnimalsRequest.java
@@ -2,9 +2,9 @@ package com.clova.anifriends.domain.animal.dto.request;
 
 import com.clova.anifriends.domain.animal.AnimalAge;
 import com.clova.anifriends.domain.animal.AnimalSize;
-import com.clova.anifriends.domain.animal.wrapper.AnimalActive;
-import com.clova.anifriends.domain.animal.wrapper.AnimalGender;
-import com.clova.anifriends.domain.animal.wrapper.AnimalType;
+import com.clova.anifriends.domain.animal.vo.AnimalActive;
+import com.clova.anifriends.domain.animal.vo.AnimalGender;
+import com.clova.anifriends.domain.animal.vo.AnimalType;
 
 public record FindAnimalsRequest(
     AnimalType type,
