+++ src/main/java/com/clova/anifriends/domain/animal/dto/response/FindAnimalByShelterResponse.java
@@ -1,9 +1,9 @@
 package com.clova.anifriends.domain.animal.dto.response;
 
 import com.clova.anifriends.domain.animal.Animal;
-import com.clova.anifriends.domain.animal.wrapper.AnimalActive;
-import com.clova.anifriends.domain.animal.wrapper.AnimalGender;
-import com.clova.anifriends.domain.animal.wrapper.AnimalType;
+import com.clova.anifriends.domain.animal.vo.AnimalActive;
+import com.clova.anifriends.domain.animal.vo.AnimalGender;
+import com.clova.anifriends.domain.animal.vo.AnimalType;
 import java.time.LocalDate;
 import java.util.List;
