+++ src/main/java/com/clova/anifriends/domain/animal/vo/AnimalAdopted.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.animal.wrapper;
+package com.clova.anifriends.domain.animal.vo;
 
 import com.clova.anifriends.domain.animal.exception.AnimalBadRequestException;
 import jakarta.persistence.Column;
