+++ src/test/java/com/clova/anifriends/domain/animal/AnimalTest.java
@@ -6,9 +6,9 @@ import static org.assertj.core.api.Assertions.catchException;
 
 import com.clova.anifriends.domain.animal.exception.AnimalBadRequestException;
 import com.clova.anifriends.domain.animal.support.fixture.AnimalFixture;
-import com.clova.anifriends.domain.animal.wrapper.AnimalActive;
-import com.clova.anifriends.domain.animal.wrapper.AnimalGender;
-import com.clova.anifriends.domain.animal.wrapper.AnimalType;
+import com.clova.anifriends.domain.animal.vo.AnimalActive;
+import com.clova.anifriends.domain.animal.vo.AnimalGender;
+import com.clova.anifriends.domain.animal.vo.AnimalType;
 import com.clova.anifriends.domain.shelter.Shelter;
 import com.clova.anifriends.domain.shelter.support.ShelterFixture;
 import java.time.LocalDate;
